refactor(live-preview): rename plugins/ into block-fields/ + media/
Context
-------
After the inline refactor, `plugins/` held 10 block-field StateFields
(table, callout, code-block, frontmatter, queryjs, mermaid,
meta-bind-button, block-math, block-comment, collection-block) plus
two media plugins (audio, video). Three semantically distinct
categories under one vague name. "Plugin" used to mean "inline plugin"
when the directory was created; today nothing inline lives there.

Split by responsibility:
  - extensions/live-preview/inline/          (unified inline pipeline)
  - extensions/live-preview/block-fields/    (ViewPlugins that replace
                                              fenced / long-block
                                              markdown with widgets)
  - extensions/live-preview/media/           (audio + video inline
                                              replacers — distinct from
                                              block fields because they
                                              replace a single line)

Solution
--------
`git mv` each file to its new folder preserving history. Updated
every import site with sed:
  - `live-preview/plugins/<name>-field.ts` → `live-preview/block-fields/<same>.ts`
  - `live-preview/plugins/<name>-plugin.ts` → `live-preview/media/<same>.ts`
  (audio-plugin.ts, video-plugin.ts)
Internal `./plugins/X` imports in live-preview.ts rewritten to the
new folders.

The original `plugins/` directory is gone; its former contents live
under two focused categories.

Testing
-------
Full suite: 5057 tests green. pnpm check clean.

Files
-----
22 moves (11 source files + 11 test files). Imports updated in:
  - src/lib/core/markdown-editor/extensions/live-preview/live-preview.ts
  - src/tests/lib/core/markdown-editor/extensions/live-preview/block-fields/*.test.ts
  - src/tests/lib/core/markdown-editor/extensions/live-preview/media/*.test.ts
  - (plus any unit tests that imported specific plugins, all absolute-path)
