+++ src/lib/core/markdown-editor/extensions/live-preview/block-fields/code-block-field.ts
