+++ src/lib/core/markdown-editor/extensions/live-preview/block-fields/collection-block-field.ts
