+++ src/lib/core/markdown-editor/extensions/live-preview/block-fields/mermaid-field.ts
