+++ src/lib/core/markdown-editor/extensions/live-preview/live-preview.ts
@@ -3,18 +3,18 @@ import type { Extension } from '@codemirror/state';
 import { livePreviewClickHandler } from './click-handler';
 import { livePreviewStyles } from './styles';
 import { mouseSelectingField, mouseSelectingHandlers } from './core/mouse-selecting';
-import { frontmatterField, frontmatterGutter } from './plugins/frontmatter-field';
-import { codeBlockField } from './plugins/code-block-field';
-import { blockCommentField } from './plugins/block-comment-field';
-import { tableField } from './plugins/table-field';
-import { calloutField } from './plugins/callout-field';
-import { collectionBlockField } from './plugins/collection-block-field';
-import { metaBindButtonField } from './plugins/meta-bind-button-field';
-import { queryjsBlockField } from './plugins/queryjs-block-field';
-import { mermaidField } from './plugins/mermaid-field';
-import { blockMathField } from './plugins/block-math-field';
-import { audioPlugin } from './plugins/audio-plugin';
-import { videoPlugin } from './plugins/video-plugin';
+import { frontmatterField, frontmatterGutter } from './block-fields/frontmatter-field';
+import { codeBlockField } from './block-fields/code-block-field';
+import { blockCommentField } from './block-fields/block-comment-field';
+import { tableField } from './block-fields/table-field';
+import { calloutField } from './block-fields/callout-field';
+import { collectionBlockField } from './block-fields/collection-block-field';
+import { metaBindButtonField } from './block-fields/meta-bind-button-field';
+import { queryjsBlockField } from './block-fields/queryjs-block-field';
+import { mermaidField } from './block-fields/mermaid-field';
+import { blockMathField } from './block-fields/block-math-field';
+import { audioPlugin } from './media/audio-plugin';
+import { videoPlugin } from './media/video-plugin';
 import { scrollDebouncePlugin } from './core/scroll-debounce-plugin';
 import { pasteTsvHandler } from './core/paste-tsv-handler';
 import { newInlineExtensions } from './inline/inline-extensions';
