+++ src/tests/lib/core/markdown-editor/extensions/live-preview/block-fields/collection-block-field.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import { EditorState, EditorSelection } from '@codemirror/state';
-import { computeCollectionBlocks } from '$lib/core/markdown-editor/extensions/live-preview/plugins/collection-block-field';
+import { computeCollectionBlocks } from '$lib/core/markdown-editor/extensions/live-preview/block-fields/collection-block-field';
 
 function createState(doc: string, cursor?: number): EditorState {
 	return EditorState.create({
