+++ src/tests/lib/core/markdown-editor/extensions/live-preview/block-fields/frontmatter-field.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import { EditorState, EditorSelection } from '@codemirror/state';
-import { computeFrontmatter } from '$lib/core/markdown-editor/extensions/live-preview/plugins/frontmatter-field';
+import { computeFrontmatter } from '$lib/core/markdown-editor/extensions/live-preview/block-fields/frontmatter-field';
 
 function createState(doc: string, cursor?: number): EditorState {
 	return EditorState.create({
